+++ lib/ModelMixins/DiscretelyTimeVaryingMixin.ts
@@ -259,7 +259,7 @@ function DiscretelyTimeVaryingMixin<
     }
 
     /**
-     * Try to calculate a multiplier which results in a new time step every {this.multiplierDefaultDeltaStep} seconds. For example, if {this.multiplierDefaultDeltaStep = 5} it would set the `multiplier` so that a new time step (of this dataset) would appear every five seconds if the timeline is playing.
+     * Try to calculate a multiplier which results in a new time step every {this.multiplierDefaultDeltaStep} seconds. For example, if {this.multiplierDefaultDeltaStep = 5} it would set the `multiplier` so that a new time step (of this dataset) would appear every five seconds (on average) if the timeline is playing.
      */
     @computed
     get multiplier() {