+++ lib/Models/ArcGisPortalCatalogGroup.ts
@@ -4,6 +4,7 @@ import URI from "urijs";
 import isDefined from "../Core/isDefined";
 import loadJson from "../Core/loadJson";
 import TerriaError from "../Core/TerriaError";
+import AccessControlMixin from "../ModelMixins/AccessControlMixin";
 import CatalogMemberMixin from "../ModelMixins/CatalogMemberMixin";
 import GroupMixin from "../ModelMixins/GroupMixin";
 import UrlMixin from "../ModelMixins/UrlMixin";
@@ -337,6 +338,13 @@ export class ArcGisPortalStratum extends LoadableStratum(
     ) {
       this.addCatalogItemByPortalGroupsToCatalogGroup(item, arcgisDataset);
     }
+
+    if (
+      AccessControlMixin.isMixedInto(item) &&
+      arcgisDataset.access !== undefined
+    ) {
+      item.setAccessType(arcgisDataset.access);
+    }
   }
 }
 
@@ -437,6 +445,13 @@ function createGroupsByPortalGroups(arcgisPortal: ArcGisPortalStratum) {
             group.description
           );
         }
+      }
+
+      if (
+        AccessControlMixin.isMixedInto(existingGroup) &&
+        group.access !== undefined
+      ) {
+        existingGroup.setAccessType(group.access);
       }
       out.push(existingGroup);
     }