+++ CHANGES.md
@@ -3,7 +3,9 @@ Change Log
 
 ### MobX Development
 
-#### next release (8.0.0-alpha.61)
+#### next release (8.0.0-alpha.62)
+
+#### 8.0.0-alpha.61
 * New `CatalogFunctionMixin` and `CatalogFunctionJobMixin`
 * Tsified `FunctionParameters`
 * New `YourDataYourRegions` `CatalogFunctionMixin`
@@ -21,6 +23,7 @@ Change Log
 * Fixed broken chart disclaimers in shared views.
 * Fixed a bug where chart disclaimers were shown even for chart items disabled in the workbench.
 * Fixed a bug where charts with titles containing the text "lat" or "lon" were hidden from feature info panel.
+* Fixed a bug that occurred when loading config from magda. `initializationUrls` are now applied even if `group` aspect is not set
 
 #### 8.0.0-alpha.60
 * Fix WMS legend for default styles.